+++ src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/git/hooks/exceptions/HookNotFoundException.kt
@@ -1,0 +1,5 @@
+package com.github.y0ung3r.gitglobalhookslocator.git.hooks.exceptions
+
+class HookNotFoundException(fileName: String)
+    : Exception("Global hook \"$fileName\" not found")
+