+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 # Changelog
 
-## [Unreleased]
+## [0.1.0]
+### Added
+- Detect changes in the Git hooks folder
+
+### Changed
+- Make the main tool window immediately available during project reindexing
 
 ## [0.0.2]
 ### Fixed
@@ -13,4 +18,5 @@
 - Git version handling
 
 ## [0.0.1]
+### Added
 - Initial release